requirements.txt
streamlit
streamlit-gps-location
fpdf2
Pillow
pandas
